+++ src/components/EditSections/EditUserInfo/components/ExternalLinkModal/index.js
@@ -1,0 +1,1 @@
+export { default as ExternalLinkModal } from './ExternalLinkModal';