+++ src/components/EditSections/EditUserInfo/components/ProfilePicture/ProfilePicture.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 
@@ -12,17 +12,21 @@ import {
 import { useStripes } from '@folio/stripes/core';
 
 import { useProfilePicture } from '../../../../../hooks';
-import { isAValidUUID } from '../../../../util/util';
+import { isAValidURL } from '../../../../util/util';
 import profilePicThumbnail from '../../../../../../icons/profilePicThumbnail.png';
 import css from '../../EditUserInfo.css';
+import ExternalLinkModal from '../ExternalLinkModal/ExternalLinkModal';
 
-const ProfilePicture = ({ label, profilePictureLink }) => {
+const ProfilePicture = ({ profilePictureId, form }) => {
+  const [profilePictureLink, setProfilePictureLink] = useState(profilePictureId);
+  const [externalLinkModalOpen, setExternalLinkModalOpen] = useState(false);
   const intl = useIntl();
   const stripes = useStripes();
-  const isProfilePictureLinkAURL = !isAValidUUID(profilePictureLink);
   const hasProfilePicture = Boolean(profilePictureLink);
-  const { isFetching, profilePictureData } = useProfilePicture({ profilePictureId: profilePictureLink });
+  const isProfilePictureLinkAURL = hasProfilePicture && isAValidURL(profilePictureLink);
   const hasAllProfilePicturePerms = stripes.hasPerm('ui-users.profile-pictures.all');
+
+  const { isFetching, profilePictureData } = useProfilePicture({ profilePictureId });
 
   const renderProfilePic = () => {
     const profilePictureSrc = isProfilePictureLinkAURL ? profilePictureLink : 'data:;base64,' + profilePictureData;
@@ -38,13 +42,30 @@ const ProfilePicture = ({ label, profilePictureLink }) => {
     );
   };
 
+  const toggleExternalLinkModal = () => {
+    setExternalLinkModalOpen(prev => !prev);
+  };
+
+  const handleSaveExternalProfilePictureLink = (externalLink) => {
+    const { change } = form;
+    change('personal.profilePictureLink', externalLink);
+    toggleExternalLinkModal();
+    setProfilePictureLink(externalLink);
+  };
+
   const renderMenu = () => (
     <DropdownMenu
       aria-label="profile picture action menu"
       role="menu"
     >
       <Button buttonStyle="dropdownItem"><Icon icon="profile">Local File</Icon></Button>
-      <Button buttonStyle="dropdownItem"><Icon icon="external-link">External URL</Icon></Button>
+      <Button
+        data-testId="externalURL"
+        buttonStyle="dropdownItem"
+        onClick={toggleExternalLinkModal}
+      >
+        <Icon icon="external-link">External URL</Icon>
+      </Button>
       <Button buttonStyle="dropdownItem"><Icon icon="trash">Delete</Icon></Button>
     </DropdownMenu>
   );
@@ -52,7 +73,9 @@ const ProfilePicture = ({ label, profilePictureLink }) => {
   return (
     <>
       <Label tagName="div">
-        {label}
+        {
+          intl.formatMessage({ id: 'ui-users.information.profilePicture' })
+        }
       </Label>
       { renderProfilePic()}
       <br />
@@ -67,13 +90,23 @@ const ProfilePicture = ({ label, profilePictureLink }) => {
           />
         )
       }
+      {
+        externalLinkModalOpen && (
+          <ExternalLinkModal
+            open={externalLinkModalOpen}
+            onClose={toggleExternalLinkModal}
+            onSave={handleSaveExternalProfilePictureLink}
+            profilePictureLink={isProfilePictureLinkAURL ? profilePictureLink : ''}
+          />
+        )
+      }
     </>
   );
 };
 
 ProfilePicture.propTypes = {
-  label: PropTypes.node.isRequired,
-  profilePictureLink: PropTypes.string,
+  form: PropTypes.object.isRequired,
+  profilePictureId: PropTypes.string.isRequired,
 };
 
 export default ProfilePicture;